requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
opencv-python-headless>=4.9.0
pyzbar>=0.1.9
Pillow>=10.3.0
numpy>=1.26.0
